rename chat to room (#117)
