+++ src/lib/ui/RoomItem.svelte
