+++ src/lib/ui/RoomItems.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type {File} from '$lib/vocab/file/file.js';
 	import type {Member} from '$lib/vocab/member/member.js';
-	import ChatItem from '$lib/ui/ChatItem.svelte';
+	import RoomItem from '$lib/ui/RoomItem.svelte';
 
 	export let files: File[];
 	export let membersById: Map<number, Member>;
@@ -17,6 +17,6 @@
 <!-- TODO possibly remove the `ul` wrapper and change the `li`s to `div`s -->
 <ul>
 	{#each files as file (file.file_id)}
-		<ChatItem {file} member={toMember(file.actor_id)} />
+		<RoomItem {file} member={toMember(file.actor_id)} />
 	{/each}
 </ul>
